Allow intcode computer to handle longs.

+++ src/main/kotlin/com/dennismcdaid/advent2019/Day05.kt
@@ -3,7 +3,7 @@ package com.dennismcdaid.advent2019
 import com.dennismcdaid.advent2019.intcode.IntcodeComputer
 
 object Day05 {
-  fun execute(intArray: IntArray, input: Int): Int =
+  fun execute(intArray: LongArray, input: Long): Long =
     IntcodeComputer(intArray)
       .apply { addInput(input) }
       .runThenGet {
