+++ src/main/kotlin/com/dennismcdaid/advent2019/Day07.kt
@@ -1,10 +1,11 @@
 package com.dennismcdaid.advent2019
 
 import com.dennismcdaid.advent2019.intcode.IntcodeComputer
+import com.dennismcdaid.advent2019.intcode.Operation
 
 object Day07 {
 
-  fun generateAllUniqueSequences(numStack: List<Int>) : List<List<Int>> {
+  fun generateAllUniqueSequences(numStack: List<Long>) : List<List<Long>> {
     if (numStack.size == 1) return listOf(numStack)
     return numStack.flatMap { num ->
       generateAllUniqueSequences(
@@ -15,12 +16,12 @@ object Day07 {
     }
   }
 
-  fun runOnAmplifier(computer: IntcodeComputer) : Int = computer.runThenGet {
+  fun runOnAmplifier(computer: IntcodeComputer) : Long = computer.runThenGet {
     it.output.last()
   }
 
-  fun getOutputForSettings(intArray: IntArray, settings: List<Int>) : Int {
-    return settings.fold(0, { total, phaseSetting ->
+  fun getOutputForSettings(intArray: LongArray, settings: List<Long>) : Long {
+    return settings.fold(0L, { total, phaseSetting ->
       val computer = IntcodeComputer(intArray.copyOf())
         .apply {
           addInput(phaseSetting)
@@ -30,13 +31,13 @@ object Day07 {
     })
   }
 
-  fun findBestPossibleConfig(intArray: IntArray, phaseSettings: List<Int>) : Int {
+  fun findBestPossibleConfig(intArray: LongArray, phaseSettings: List<Long>) : Long {
     return generateAllUniqueSequences(phaseSettings)
       .map { getOutputForSettings(intArray, it) }
       .max() ?: throw IllegalStateException("There was no max!")
   }
 
-  private fun setupAmps(memory: IntArray, phaseSettings: List<Int>) : Iterator<Pair<Int, IntcodeComputer>> {
+  private fun setupAmps(memory: LongArray, phaseSettings: List<Long>) : Iterator<Pair<Int, IntcodeComputer>> {
     return phaseSettings.map {
       IntcodeComputer(memory.copyOf())
         .apply { addInput(it) }
@@ -45,9 +46,9 @@ object Day07 {
       .indexedCycleIterator()
   }
 
-  fun getFeedbackForSequence(memory: IntArray, phaseSettings: List<Int>) : Int {
+  fun getFeedbackForSequence(memory: LongArray, phaseSettings: List<Long>) : Long {
     val amps = setupAmps(memory, phaseSettings)
-    var input = 0
+    var input = 0L
     var ampNumber: Int
     var amp: IntcodeComputer
     do {
@@ -57,15 +58,15 @@ object Day07 {
       amp.addInput(input)
       val originalOutputSize = amp.output.size
       amp.runUntil {
-        it.output.size > originalOutputSize || it.current == 99
+        it.output.size > originalOutputSize || it.current == 99L
       }.let {
         input = it.output.last()
       }
-    } while (amp.current != 99 || ampNumber != phaseSettings.size-1)
+    } while (amp.current != 99L || ampNumber != phaseSettings.size-1)
     return input
   }
 
-  fun findBestPossibleFeedback(memory: IntArray, phaseSettings: List<Int>) : Int {
+  fun findBestPossibleFeedback(memory: LongArray, phaseSettings: List<Long>) : Long {
     return generateAllUniqueSequences(phaseSettings)
       .map { getFeedbackForSequence(memory, it) }
       .max() ?: throw IllegalStateException("There was no max!")
